Move state manager & execution context to ‘runtime’ module space
