+++ src/runtime/state_manager_test.ts
